requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-cors==4.0.2
pymongo>=4.6.3
python-dotenv==1.0.0
pdfplumber==0.10.1
requests>=2.32.0
Werkzeug==3.0.6
gunicorn==23.0.0
lmstudio==0.0.1
pyyaml
fastapi

# API tabanlı LLM'ler için bağımlılıklar
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.5.0

# Güvenlik ve şifreleme için
cryptography>=41.0.0